requirements.txt
accelerate
diffusers
invisible_watermark
torch
transformers
xformers
flask
Pillow
numpy
